chore(infra-dynamodb): move GetUserProfileByEmailRepository to its own file
