+++ src/infra/db/dynamodb/repositories/index.js
@@ -1,0 +1,5 @@
+const GetUserProfileByEmailRepository = require('./get-user-profile-by-email-repository');
+
+module.exports = {
+  GetUserProfileByEmailRepository
+};
